+++ src/projects.py
@@ -91,23 +91,23 @@ class Projects(ndb.Model):
                 lic_img = "https://i.creativecommons.org/l/by/4.0/88x31.png"
                 lic_txt = "Creative Commons Attribution 4.0 International License"
             elif lic == "CC BY-SA":
-                lic_url = "https://creativecommons.org/licenses/by/4.0/"
+                lic_url = "https://creativecommons.org/licenses/by-sa/4.0/"
                 lic_img = "https://i.creativecommons.org/l/by-sa/4.0/88x31.png"
                 lic_txt = "Creative Commons Attribution-ShareAlike 4.0 International License"
             elif lic == "CC BY-ND":
-                lic_url = "https://creativecommons.org/licenses/by/4.0/"
+                lic_url = "https://creativecommons.org/licenses/by-nd/4.0/"
                 lic_img = "https://i.creativecommons.org/l/by-nd/4.0/88x31.png"
                 lic_txt = "Creative Commons Attribution-NoDerivatives 4.0 International License"
             elif lic == "CC BY-NC":
-                lic_url = "https://creativecommons.org/licenses/by/4.0/"
+                lic_url = "https://creativecommons.org/licenses/by-nc/4.0/"
                 lic_img = "https://i.creativecommons.org/l/by-nc/4.0/88x31.png"
                 lic_txt = "Creative Commons Attribution-NonCommercial 4.0 International License"
             elif lic == "CC BY-NC-SA":
-                lic_url = "https://creativecommons.org/licenses/by/4.0/"
+                lic_url = "https://creativecommons.org/licenses/by-nc-sa/4.0/"
                 lic_img = "https://i.creativecommons.org/l/by-nc-sa/4.0/88x31.png"
                 lic_txt = "Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International License"
             elif lic == "CC BY-NC-ND":
-                lic_url = "https://creativecommons.org/licenses/by/4.0/"
+                lic_url = "https://creativecommons.org/licenses/by-nc-nd/4.0/"
                 lic_img = "https://i.creativecommons.org/l/by-nc-nd/4.0/88x31.png"
                 lic_txt = "Creative Commons Attribution-NonCommercial-NoDerivatives 4.0 International License"
             return '<a rel="license" href="%s" class="license-img"><img alt="Creative Commons License" style="border-width:0" src="%s" /></a><div class="license-text">Unless otherwise specified, original content from <b><span xmlns:dct="https://purl.org/dc/terms/" property="dct:title">%s</span></b> is licensed under a<br><a rel="license" href="%s">%s</a>.</div>' % (lic_url, lic_img, self.name, lic_url, lic_txt)